// client/src/types/types.ts
export type User = {
  name: string;
  email: string;
  imgUrl: string;
  gender: string;
  role: string;
  dob: string;
  _id: string;
};

export type Product = {
  name: string;
  price: number;
  stock: number;
  category: string;
  imgUrl: string;
  _id: string;
};

export type ShippingInfo = {
  address: string;
  city: string;
  state: string;
  country: string;
  pincode: string;
};

export type CartItem = {
  productId: string;
  imgUrl: string;
  name: string;
  price: number;
  quantity: number;
  stock: number;
};

export type OrderItem = {
  productId: string;
  imgUrl: string;
  name: string;
  price: number;
  quantity: number;
  _id: string;
};

export type Order = {
  orderItems: OrderItem[];
  shippingInfo: ShippingInfo;
  subtotal: number;
  tax: number;
  shippingCharges: number;
  discount: number;
  total: number;
  status: string;
  user: {
    name: string;
    _id: string;
  };
  _id: string;
};